#249 updated script to log the number of updated records and upda operations performed
